% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSamples.R
\name{analysis.samples}
\alias{analysis.samples}
\title{Library size and sample wise threshold analysis.}
\usage{
analysis.samples(object)
}
\arguments{
\item{object}{an RTIGER object}
}
\value{
A list with the library size of ech sample as total allele count, and a logical vector which specifies wether or not each samples pass the threshold set when the object was created.
}
\description{
Library size and sample wise threshold analysis.
}
\examples{

data("fittedExample")
myAnalysis = analysis.samples(myDat)

}
