% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATseqlengths}
\alias{ATseqlengths}
\title{The autosome chromosome lengths for Arabidopsis Thaliana.}
\description{
The autosome chromosome lengths for Arabidopsis Thaliana.
}
\author{
Rafael Campos-Martin
}
\keyword{data}
