% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO-analysis.R
\name{calcDoubleCOnumber}
\alias{calcDoubleCOnumber}
\title{Compute number of double CO per sample and chromosome.}
\usage{
calcDoubleCOnumber(object, samples = NULL)
}
\arguments{
\item{object}{a RViterbi object.}

\item{samples}{character vector with the name of the samples.}
}
\description{
Compute number of double CO per sample and chromosome.
}
