% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conf_Interval.R
\name{conf.intervals}
\alias{conf.intervals}
\title{Obtain a list of Genomic Ranges object with the confidence interval (genomic Region), which a CO over has happened above certain threshold conf.}
\usage{
conf.intervals(rtigerobj, rviterbiobj, conf = 0.95, max.width = 3, samples = NULL)
}
\arguments{
\item{rtigerobj}{a RTiger object.}

\item{rviterbiobj}{a RViterbi object.}

\item{conf}{value between 0 and 1. Indicates the percentage of the confidence interval.}

\item{max.width}{maximum number of observations apart to look for the probability.}

\item{samples}{If wanted. It can be applied ot a smaller set of samples.}
}
\value{
List of genomic Ranges object
}
\description{
Obtain a list of Genomic Ranges object with the confidence interval (genomic Region), which a CO over has happened above certain threshold conf.
}
\examples{

data("fittedExample")
data("R.ViterbiExample")
my.conf.int = conf.intervals(myDat, RVit)

}
