% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developer.R
\name{dev}
\alias{dev}
\title{Function to developers. It runs one EM step}
\usage{
dev(psi, rigidity = NULL, nstates = 3, transition = NULL, start = NULL)
}
\arguments{
\item{psi}{list of psi probabilities.}

\item{rigidity}{Rigidity value.}

\item{nstates}{Number of states.}

\item{transition}{transition matrix}

\item{start}{initial probabilities}
}
\value{
List with updates probabilites
}
\description{
Function to developers. It runs one EM step
}
