% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Package.R
\docType{data}
\name{expDesign}
\alias{expDesign}
\title{data frame with file address and names}
\description{
data frame with file address and names
}
\author{
Rafael Campos-Martin, Laia Capilla, Victor Solier, Raphael Mercier, Korbinian Schneeberger, Achim Tresch
}
\keyword{data}
