% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFit.R
\name{fitModel}
\alias{fitModel}
\title{This function fits our extended HMM to an object with the observations}
\usage{
fitModel(object, initial_params = NULL, randomize = FALSE,
  eps = 0.01, max.iter = 50, trace = FALSE, verbose = FALSE)
}
\arguments{
\item{object}{an RTIGER object with the filtered data}

\item{initial_params}{initial params to start the EM algorithm}

\item{randomize}{Logical value, wehter to randomize the initial parameters for the EM algorithm}

\item{eps}{convergence cutoff for EM algorithm. Difference of the the parameters between two consecutive iterations.}

\item{max.iter}{Maximum number of iterations.}

\item{trace}{Logical value, wehter the values of the parameters along the EM algorithm should be saved}

\item{verbose}{logical value, wether the algorithm status should be printed or not.}
}
\description{
This function fits our extended HMM to an object with the observations
}
