% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myDat}
\alias{myDat}
\title{A fitted example using three own samples of Arabidopsis. More information in publication:}
\description{
A fitted example using three own samples of Arabidopsis. More information in publication:
}
\author{
Rafael Campos-Martin
}
\keyword{data}
