% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCO.R
\name{plotCOs}
\alias{plotCOs}
\title{Obtain number of Cross-Over events per sample and chromosome.}
\usage{
plotCOs(object, file = NULL)
}
\arguments{
\item{object}{a RViterbi object.}

\item{file}{file where to save the plot for CO numbers}
}
\value{
a plot
}
\description{
Obtain number of Cross-Over events per sample and chromosome.
}
\examples{

data("fittedExample")
co.num = calcCOnumber(myDat)

}
