% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotGenotype}
\alias{plotGenotype}
\title{This funciton creates an RViterbi object.}
\usage{
plotGenotype(object, samp, chr, col = c("red", "blue", "mediumorchid4"), size = c(1,1), cex.feature = .7, from = NULL, to = NULL, main = NULL, showId = TRUE, showGenAxis = FALSE)
}
\arguments{
\item{object}{an RTIGER or RViterbi object with the filtered data}

\item{samp}{character with the sample name as specified in the experimentalDesign}

\item{chr}{character with the chromosome to plot}

\item{col}{color to define segments in P1, P2 and heterozygous regions.}

\item{size}{size of each  panel (see Gviz documentation)}

\item{...}{Additional arguments which are all interpreted as display parameters to tweak the appearance of the plot. These parameters are global, meaning that they will be used for all tracks in the list where they actually make sense, and they override the track-internal settings. See Gviz, plotTracks funciton for details on display parameters.}
}
\description{
This funciton creates an RViterbi object.
}
\examples{

data("R.ViterbiExample")
info = myDat@info
plotGenotype(R.Vit, samp = info$sample_names[1], info$part_names[1])
}
