% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RViterbi.R
\name{R.Viterbi}
\alias{R.Viterbi}
\title{This funciton creates an RViterbi object.}
\usage{
R.Viterbi(object, Rs)
}
\arguments{
\item{object}{an RTIGER object with the filtered data}

\item{Rs}{rigidity values to be checked. Number of consecutive iterations in state s before using the transition matrix.}
}
\value{
returns and RViterbi object with the new viterbi path with rigidity R.
}
\description{
This funciton creates an RViterbi object.
}
\examples{

data("fittedExample")
RVit = R.Viterbi(myDat, Rs = 3:4)

}
