% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawImport.R
\name{RawDataSetImport}
\alias{RawDataSetImport}
\title{This function creates an RTIGER Object with raw reads}
\usage{
RawSetImport(experimentDesign = NULL, observations = NULL, GenRanges = NULL,min.samples = 1, seqlengths = NULL)
}
\arguments{
\item{experimentDesign}{Data frame with minimum two columns: files and name}

\item{observations}{A list of samples, which each sample is a list of matrix. Each matrix is a 2 by N markers. The first row is the parent 1 allele count and the second row is the total allele count.}

\item{GenRanges}{A GenomicRanges object with the positions of each marker.}

\item{min.samples}{Minumum number of samples supporting an observation to be kept. Default is 1 sample.}

\item{seqlengths}{The genomic lengths of each chromosome (As in GenomicRanges).}
}
\value{
A RTiger object with Genomic Ranges for the filtered and Raw data
}
\description{
This function creates an RTIGER Object with raw reads
}
\examples{

data("ATseqlengths")
path = system.file("extdata",  package = "RTIGER")
files = list.files(path, full.names = TRUE)[1:3]
expDesign = data.frame(files = files, name = list.files(path)[1:3])
myDat = RawDataSetImport(experimentDesign = expDesign, bin.length = 100, seqlengths = ATseqlengths[1:4])


}
