% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Package.R
\docType{package}
\name{RTIGER-package}
\alias{RTIGER-package}
\title{RTIGER: A Robust HMM-based Model for genotyping-by-sequencing and Recombination breakpoint identification.}
\description{
RTIGER package
}
\author{
Rafael Campos-Martin, Korbinian Schneeberger, Achim Tresch.
}
\keyword{package}
