% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Package.R
\docType{data}
\name{RVit}
\alias{RVit}
\title{fittedExample after applying the R.Viterbi algorithm}
\description{
fittedExample after applying the R.Viterbi algorithm
}
\author{
Rafael Campos-Martin, Laia Capilla, Victor Solier, Raphael Mercier, Korbinian Schneeberger, Achim Tresch
}
\keyword{data}
