% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RViterbi.R
\docType{class}
\name{RViterbi-class}
\alias{RViterbi-class}
\alias{.RViterbi}
\title{This class is a generic container for R-Viterbi. It contains the modified decoding version of the Viterbi algorithm}
\description{
This class is a generic container for R-Viterbi. It contains the modified decoding version of the Viterbi algorithm
}
\section{Slots}{

\describe{
\item{\code{Rparameter}}{The R parameter. It specifies the minimum number of iterations over a state before a transition is permited.}

\item{\code{LogLike}}{loglikelihood of the observed viterbi path given the observations.}

\item{\code{Viterbi}}{List of chromosomes with the viterbi path per sample}
}}

