% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupJulia.R
\name{setupJulia}
\alias{setupJulia}
\title{Installs the needed packages in JULIA to run the EM algorithm for rHMM.}
\usage{
setupJulia(JULIA_HOME = NULL)
}
\arguments{
\item{JULIA_HOME}{the file folder which contains julia binary, if not set, JuliaCall will look at the global option JULIA_HOME, if the global option is not set, JuliaCall will then look at the environmental variable JULIA_HOME, if still not found, JuliaCall will try to use the julia in path.}
}
\value{
empty
}
\description{
Installs the needed packages in JULIA to run the EM algorithm for rHMM.
}
