% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2IGV.R
\name{snpCounts2IGV}
\alias{snpCounts2IGV}
\title{Exports the count at the marker level to a bw file to visualize in IGV}
\usage{
snpCounts2IGV(object, sample, dir = NULL)
}
\arguments{
\item{object}{an RTIGER object with the filtered data}

\item{sample}{character with the sample name as specified in the experimentalDesign}

\item{dir}{character with the chromosome to plot. If left null, it will create a folder with the name of the sample.}
}
\value{
a folder with files to visualize results in igv.
}
\description{
Exports the count at the marker level to a bw file to visualize in IGV
}
\examples{

data("fittedExample.rda")
info = myDat@info
snpCounts2IGV(myDat, samp = info$sample_names[1])
}
