% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupJulia.R
\name{sourceJulia}
\alias{sourceJulia}
\title{Function needed before using RTIGER() function. It loads the scripts in Julia that fit the rHMM.}
\usage{
sourceJulia()
}
\value{
empty
}
\description{
Function needed before using RTIGER() function. It loads the scripts in Julia that fit the rHMM.
}
