% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_ratios.R
\name{state_ratios}
\alias{state_ratios}
\title{This funciton creates matrix with the state count ratio}
\usage{
state_ratios(RVit)
}
\arguments{
\item{rviterbiobj}{an RVITERBI object.}

\item{samples}{the samples used to compute the ratio}
}
\value{
returns a nx3 matrix with n being the number of samples
}
\description{
This funciton creates matrix with the state count ratio
}
\examples{

data("R.ViterbiExample")
ratios = state_ratios(RVit)

}
