% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO-analysis.R
\name{width.DoubleCO}
\alias{width.DoubleCO}
\title{genomic widths of the double COs.}
\usage{
width.DoubleCO(object, samples = NULL)
}
\arguments{
\item{object}{a RViterbi object.}

\item{samples}{character vector with the name of the samples.}
}
\description{
genomic widths of the double COs.
}
