% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.csv.R
\name{write.geno.csv}
\alias{write.geno.csv}
\title{write.geno.csv prints its required Viterbi path from an object (after converting it to a data frame if it is not one nor a matrix) to a file or connection.}
\usage{
write.geno.csv(object, file, samples = NULL, chr = NULL)
}
\arguments{
\item{object}{A Rtiger or RViterbi object}

\item{file}{ither a character string naming a file or a connection open for writing. "" indicates output to the console.}

\item{samples}{a set of samples. If null, all samples will be used}

\item{chr}{The Chromosome to use.}
}
\value{
List of genomic Ranges object
}
\description{
write.geno.csv prints its required Viterbi path from an object (after converting it to a data frame if it is not one nor a matrix) to a file or connection.
}
\examples{

data("R.ViterbiExample")
write.geno.csv(RVit, file = "example.geno.csv", chr = 1)

}
