% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Package.R
\docType{data}
\name{ATseqlengths}
\alias{ATseqlengths}
\title{Genomic lengths of the Arabidopsis Thalian chromosomes.}
\description{
Genomic lengths of the Arabidopsis Thalian chromosomes.
}
\author{
Rafael Campos-Martin, Laia Capilla, Victor Solier, Raphael Mercier, Korbinian Schneeberger, Achim Tresch
}
\keyword{data}
