% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO-analysis.R
\name{calcCOnumber}
\alias{calcCOnumber}
\title{Obtain number of Cross-Over events.}
\usage{
calcCOnumber(object)
}
\arguments{
\item{object}{a RViterbi object.}
}
\description{
Obtain number of Cross-Over events.
}
