% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DataSetImportFromtxt}
\alias{DataSetImportFromtxt}
\title{This function creates an RTIGER Object}
\usage{
DataSetImportFromtxt(experimentDesign = NULL, observations = NULL, GenRanges = NULL, bin.length = NULL, min.samples = 1, min.counts = 10, quant = .2, seqlengths = NULL)
}
\arguments{
\item{experimentDesign}{Data frame with minimum two columns: files and name}

\item{observations}{A list of samples, which each sample is a list of matrix. Each matrix is a 2 by N markers. The first row is the parent 1 allele count and the second row is the total allele count.}

\item{GenRanges}{A GenomicRanges object with the positions of each marker.}

\item{bin.length}{Length for which start binning the genome and by which will be increased on each iteration}

\item{min.samples}{Minumum number of samples supporting an observation to be kept. Default is 1 sample.}

\item{min.counts}{Total allele count to select a criterion to keep increasing the bin size.}

\item{quant}{Quantile of observations that support the minimum counts. This quantile is used to decide when to stop increasing the bin size}

\item{seqlengths}{The genomic lengths of each chromosome (As in GenomicRanges).}
}
\value{
A RTiger object with Genomic Ranges for the filtered and Raw data
}
\description{
This function creates an RTIGER Object
}
\examples{

data("ATseqlengths")
path = system.file("extdata",  package = "RTIGER")
files = list.files(path, full.names = TRUE)[1:3]
expDesign = data.frame(files = files, name = list.files(path)[1:3])
myDat = DataSetImportFromtxt(experimentDesign = expDesign, bin.length = 100, seqlengths = ATseqlengths)


}
