% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2IGV.R
\name{export2IGV}
\alias{export2IGV}
\title{Exports the count and viterbi code to a folder in order to visualize it in IGV.}
\usage{
export2IGV(object, sample, dir = NULL, ratio = FALSE)
}
\arguments{
\item{object}{an RTIGER or RViterbi object with the filtered data}

\item{sample}{character with the sample name as specified in the experimentalDesign}

\item{dir}{character with the chromosome to plot. If left null, it will create a folder with the name of the sample.}

\item{ratio}{boolean variable. Whether the ratio P1/total should be created as well.}
}
\value{
a folder with files to visualize results in igv.
}
\description{
Exports the count and viterbi code to a folder in order to visualize it in IGV.
}
\examples{

data("R.ViterbiExample")
info = myDat@info
export2IGV(RVit, samp = info$sample_names[1])
}
