% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO-analysis.R
\name{haplotypes.width}
\alias{haplotypes.width}
\title{Haplotypes genomic widths.}
\usage{
haplotypes.width(object)
}
\arguments{
\item{object}{a RViterbi object.}
}
\description{
Haplotypes genomic widths.
}
