% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Package.R
\docType{data}
\name{myDat}
\alias{myDat}
\title{Fitted RTiger object from 3 samples}
\description{
Fitted RTiger object from 3 samples
}
\author{
Rafael Campos-Martin, Laia Capilla, Victor Solier, Raphael Mercier, Korbinian Schneeberger, Achim Tresch
}
\keyword{data}
