% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenotype.R
\name{plotGenotype}
\alias{plotGenotype}
\title{This funciton creates an RViterbi object.}
\usage{
plotGenotype(object, samp, chr, col = c("red", "blue", "mediumorchid4"), size = c(1,1), main = NULL, showGenAxis = TRUE, ratio = FALSE, window = 10)
}
\arguments{
\item{object}{an RTIGER or RViterbi object with the filtered data}

\item{samp}{character with the sample name as specified in the experimentalDesign}

\item{chr}{character with the chromosome to plot}

\item{col}{color to define segments in P1, P2 and heterozygous regions.}

\item{size}{size of each  panel (see Gviz documentation)}

\item{main}{an overall title for the plot. If null, the sample and chromosome are the default.}

\item{showGenAxis}{boolean parameter whether or not show the genome axis.}

\item{ratio}{boolean parameter wether or not to plot the ratio counts}

\item{window}{if ratio TRUE, what is the window for the sliding window to be computed of the ratios.}
}
\value{
plot with the genomic segmentation for chromosome chr in sample samp.
}
\description{
This funciton creates an RViterbi object.
}
\examples{

data("fittedExample")
info = myDat@info
plotGenotype(myDat, samp = info$sample_names[1], info$part_names[1])
}
