% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Class.R
\docType{class}
\name{RTIGER-class}
\alias{RTIGER-class}
\alias{.RTIGER}
\title{This class is a generic container for RTIGER analysis}
\description{
This class is a generic container for RTIGER analysis
}
\section{Slots}{

\describe{
\item{\code{matobs}}{Nested lists. the first level is a list of samples. For each sample there are 5 matrices that contains the allele counts for each position.}

\item{\code{params}}{a list with the parameters after training.}

\item{\code{info}}{List with phenotipic data of the samples.}

\item{\code{Viterbi}}{List of chromosomes with the viterbi path per sample.}

\item{\code{Probabilities}}{Computed probabilites for the EM algorithm.}

\item{\code{num.iter}}{Number of iterations needed to stop the EM algorithm.}
}}

