% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTIGER-Class.R
\docType{class}
\name{RTIGER-class}
\alias{RTIGER-class}
\alias{.RTIGER}
\title{This class is a generic container for RTIGER analysis}
\description{
This class is a generic container for RTIGER analysis
}
\section{Slots}{

\describe{
\item{\code{RawData}}{Raw data}

\item{\code{FilteredData}}{List containing the filtered data based on the minimum number of samples supporting the each position and summing over the bin size.}

\item{\code{FilteringThreshold}}{List of the parameters used to decide the bin size.}

\item{\code{info}}{List with information about the samples.}

\item{\code{Model}}{fitted model to the data.}

\item{\code{Viterbi}}{List of chromosomes with the viterbi path per sample.}
}}

