% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupJulia.R
\name{setupJulia}
\alias{setupJulia}
\title{Installs the needed packages in JULIA to run the EM algorithm for rHMM.}
\usage{
setupJulia()
}
\description{
Installs the needed packages in JULIA to run the EM algorithm for rHMM.
}
