% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_ratios.R
\name{state_ratios}
\alias{state_ratios}
\title{This funciton computes the mean raatio per state and sample}
\usage{
state_ratios(rtigerobj)
}
\arguments{
\item{rtigerobj}{an RTIGER or RViterbi object with the filtered data}
}
\value{
matrix with the mean count ratios per sample and state.
}
\description{
This funciton computes the mean raatio per state and sample
}
\examples{
\dontrun{
data("fittedExample")
myratios = state_ratios(myDat)
}

}
